Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

FletcherPenaltyNLPSolver.FletcherPenaltyNLP

FletcherPenaltyNLPSolver.feasibility_step

FletcherPenaltyNLPSolver.fps_solve

FletcherPenaltyNLPSolver.status_stopping_to_stats

FletcherPenaltyNLPSolver.stopping_to_stats

NLPModelsIpopt.ipopt

​

FletcherPenaltyNLPSolver.FletcherPenaltyNLP — Type

We consider here the implementation of Fletcher's exact penalty method for the minimization problem:

minₓ f(x) s.t. c(x) = 0

using Fletcher penalty function:

minₓ f(x) - dot(c(x),ys(x)) + ρ/2 dot(c(x),c(x))

where

ys(x) := argmin\_y 0.5 ||A(x)y - g(x)||²₂ + σ c(x)^T y + 0.5 δ ||²₂

and denote Ys the gradient of ys(x).

FletcherPenaltyNLP(:: AbstractNLPModel, :: Number, :: Function) or FletcherPenaltyNLP(:: AbstractNLPModel; σ_0 :: Real = 1.0)

Notes:

Evaluation of the obj, grad, objgrad functions evaluate functions from the orginial nlp.

These values are stored in fx, cx, gx.

The value of the penalty vector ys is also stored.

TODO:

sparse structure of the hessian?

Example: fpsos = FletcherPenaltyNLP(nlp, 0.1, _solvewithlinearoperator)

FletcherPenaltyNLPSolver.feasibility_step — Method

feasibility_step(nls, x, cx, Jx)

Approximately solves min ‖c(x)‖.

Given xₖ, finds min ‖cₖ + Jₖd‖

FletcherPenaltyNLPSolver.fps_solve — Method

Solver for equality constrained non-linear programs based on Fletcher's penalty function.

Cite: Estrin, R., Friedlander, M. P., Orban, D., & Saunders, M. A. (2020).
Implementing a smooth exact penalty function for equality-constrained nonlinear optimization.
SIAM Journal on Scientific Computing, 42(3), A1809-A1835.

fps_solve(:: NLPStopping, :: AbstractVector{T}; σ_0 :: Number = one(T), σ_max :: Number = 1/eps(T), σ_update :: Number = T(1.15), unconstrained_solver :: Function = lbfgs) where T <: AbstractFloat or fps_solve(:: AbstractNLPModel, :: AbstractVector{T}, σ_0 :: Number = one(T), σ_max :: Number = 1/eps(T), σ_update :: Number = T(1.15), unconstrained_solver :: Function = lbfgs) where T <: AbstractFloat

Notes:

If the problem has inequalities, we use slack variables to get only equalities and bounds.

stp.current_state.res contains the gradient of Fletcher's penalty function.

unconstrained_solver must take an NLPStopping as input.

TODO:

une façon robuste de mettre à jour le paramètre de pénalité. [Convergence to infeasible stationary points]

Extend to bounds and inequality constraints.

Handle the tol_check from the paper !

Continue to explore the paper.

[Long term] Complemetarity constraints

FletcherPenaltyNLPSolver.status_stopping_to_stats — Method

Return the status in GenericStats from a Stopping.

FletcherPenaltyNLPSolver.stopping_to_stats — Method

Initialize a GenericStats from Stopping

NLPModelsIpopt.ipopt — Method

ipopt(nlp) DOESN'T CHECK THE WRONG KWARGS, AND RETURN AN ERROR. ipopt(::NLPStopping)